refactor: 로그 수정
